authentication

See merge request Blaker/ynov-dizifymusic-api!10

+++ src/main/java/projet/ynov/dizifymusicapi/controllers/FavoriteController.java
@@ -1,5 +1,6 @@
 package projet.ynov.dizifymusicapi.controllers;
 
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
@@ -8,6 +9,8 @@ import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -23,7 +26,7 @@ import projet.ynov.dizifymusicapi.entity.Favorite;
 import projet.ynov.dizifymusicapi.entity.Title;
 import projet.ynov.dizifymusicapi.entity.User;
 import projet.ynov.dizifymusicapi.entity.params.FavoriteParams;
-import projet.ynov.dizifymusicapi.exceptions.ResourceNotFoundException;
+import projet.ynov.dizifymusicapi.exceptions.GlobalHttpException;
 import projet.ynov.dizifymusicapi.repositories.AlbumRepository;
 import projet.ynov.dizifymusicapi.repositories.ArtistRepository;
 import projet.ynov.dizifymusicapi.repositories.FavoriteRepository;
@@ -45,6 +48,19 @@ public class FavoriteController {
 	@Autowired
 	private UserRepository userRepository;
 	
+	private User getUserLogged() {
+		Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		String username;
+		
+		if (principal instanceof UserDetails) {
+			username = ((UserDetails)principal).getUsername();
+		} else {
+			username = principal.toString();
+		}
+		
+		return userRepository.findByUsername(username);
+	}
+	
 	
 	/**
 	 * Get all Favorite list.
@@ -53,7 +69,7 @@ public class FavoriteController {
 	 */
 	@GetMapping("/favorites")
 	public List<Favorite> getAllFavorites() {
-		return favoriteRepository.findAll();
+		return new ArrayList<Favorite>(getUserLogged().getFavorites());
     }
 
 	/**
@@ -61,15 +77,22 @@ public class FavoriteController {
 	 *
 	 * @param FavoriteId the Favorite id
 	 * @return the Favorites by id
-	 * @throws ResourceNotFoundException the resource not found exception
+	 * @throws GlobalHttpException the resource not found exception
 	 */
 	@GetMapping("/favorites/{id}")
-	public ResponseEntity<Favorite> getFavoritesById(@PathVariable(value = "id") Long favoriteId) throws ResourceNotFoundException {
+	public ResponseEntity<Favorite> getFavoritesById(@PathVariable(value = "id") Long favoriteId) throws GlobalHttpException {
+		
+		User userLogged = getUserLogged();
+
 		Favorite favorite = favoriteRepository
 			  				.findById(favoriteId)
-	  						.orElseThrow(() -> new ResourceNotFoundException(HttpStatus.NOT_FOUND, "Favorite not found with id : " + favoriteId));
-	  
-		return ResponseEntity.ok().body(favorite);
+	  						.orElseThrow(() -> new GlobalHttpException(HttpStatus.NOT_FOUND, "Favorite not found with id : " + favoriteId));
+		
+		if (favorite.getUser().getId() == userLogged.getId()) {			
+			return ResponseEntity.ok().body(favorite);
+		}
+		
+		throw new GlobalHttpException(HttpStatus.FORBIDDEN, "Not authorized");
 	}
 
 	/**
@@ -81,32 +104,30 @@ public class FavoriteController {
 	 */
 	@PostMapping("/favorites")
 	public Favorite createFavorite(@Validated @RequestBody FavoriteParams params) throws Exception {
-		User user = userRepository
-			  				.findById(params.getUser_id())
-	  						.orElseThrow(() -> new ResourceNotFoundException(HttpStatus.NOT_FOUND, "User not found with id : " + params.getUser_id()));
+		User userLogged = getUserLogged();
 		
 		params.setCreatedAt(new Date());
 		params.setUpdatedAt(new Date());
 
 		Favorite favorite = new Favorite(params);
-		favorite.setUser(user);
+		favorite.setUser(userLogged);
 		
 		if (params.getAlbum_id() != 0L) {
 			Album album = albumRepository
 			  					.findById(params.getAlbum_id())
-								.orElseThrow(() -> new ResourceNotFoundException(HttpStatus.NOT_FOUND, "Album not found with id : " + params.getAlbum_id()));
+								.orElseThrow(() -> new GlobalHttpException(HttpStatus.NOT_FOUND, "Album not found with id : " + params.getAlbum_id()));
 			
 			favorite.setAlbum(album);
 		} else if (params.getArtist_id() != 0L) {
 			Artist artist = artistRepository
 			  					.findById(params.getArtist_id())
-								.orElseThrow(() -> new ResourceNotFoundException(HttpStatus.NOT_FOUND, "Artist not found with id : " + params.getArtist_id()));
+								.orElseThrow(() -> new GlobalHttpException(HttpStatus.NOT_FOUND, "Artist not found with id : " + params.getArtist_id()));
 
 			favorite.setArtist(artist);
 		} else if (params.getTitle_id() != 0L) {
 			Title title = titleRepository
 			  					.findById(params.getTitle_id())
-								.orElseThrow(() -> new ResourceNotFoundException(HttpStatus.NOT_FOUND, "Title not found with id : " + params.getTitle_id()));
+								.orElseThrow(() -> new GlobalHttpException(HttpStatus.NOT_FOUND, "Title not found with id : " + params.getTitle_id()));
 
 			favorite.setTitle(title);
 		} else {
@@ -126,13 +147,20 @@ public class FavoriteController {
 	 */
 	@DeleteMapping("/favorites/{id}")
 	public Map<String, Boolean> deleteFavorite(@PathVariable(value = "id") Long favoriteId) throws Exception {
-	    Favorite favorite = favoriteRepository
-	            			.findById(favoriteId)
-	            			.orElseThrow(() -> new ResourceNotFoundException(HttpStatus.NOT_FOUND, "Favorite not found with id : " + favoriteId));
 
-	    favoriteRepository.delete(favorite);
-	    Map<String, Boolean> response = new HashMap<>();
-	    response.put("deleted", Boolean.TRUE);
-	    return response;
+		User userLogged = getUserLogged();
+		
+	    Favorite favorite = favoriteRepository
+    			.findById(favoriteId)
+    			.orElseThrow(() -> new GlobalHttpException(HttpStatus.NOT_FOUND, "Favorite not found with id : " + favoriteId));
+		
+		if (favorite.getUser().getId() == userLogged.getId()) {
+		    favoriteRepository.delete(favorite);
+		    Map<String, Boolean> response = new HashMap<>();
+		    response.put("deleted", Boolean.TRUE);
+		    return response;
+		}
+		
+		throw new GlobalHttpException(HttpStatus.FORBIDDEN, "Not authorized");
 	}
 }
