+++ src/main/java/projet/ynov/dizifymusicapi/controllers/PlaylistController.java
@@ -1,5 +1,6 @@
 package projet.ynov.dizifymusicapi.controllers;
 
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -9,6 +10,8 @@ import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -23,7 +26,7 @@ import projet.ynov.dizifymusicapi.entity.Playlist;
 import projet.ynov.dizifymusicapi.entity.Title;
 import projet.ynov.dizifymusicapi.entity.User;
 import projet.ynov.dizifymusicapi.entity.params.PlaylistParams;
-import projet.ynov.dizifymusicapi.exceptions.ResourceNotFoundException;
+import projet.ynov.dizifymusicapi.exceptions.GlobalHttpException;
 import projet.ynov.dizifymusicapi.repositories.PlaylistRepository;
 import projet.ynov.dizifymusicapi.repositories.TitleRepository;
 import projet.ynov.dizifymusicapi.repositories.UserRepository;
@@ -39,6 +42,19 @@ public class PlaylistController {
 	@Autowired
 	private TitleRepository titleRepository;
 	
+	private User getUserLogged() {
+		Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		String username;
+		
+		if (principal instanceof UserDetails) {
+			username = ((UserDetails)principal).getUsername();
+		} else {
+			username = principal.toString();
+		}
+		
+		return userRepository.findByUsername(username);
+	}
+	
 	/**
 	 * Get all Playlist list.
 	 *
@@ -46,7 +62,7 @@ public class PlaylistController {
 	 */
 	@GetMapping("/playlists")
 	public List<Playlist> getAllPlaylists() {
-		return playlistRepository.findAll();
+		return new ArrayList<Playlist>(getUserLogged().getPlaylists());
     }
 
 	/**
@@ -54,15 +70,21 @@ public class PlaylistController {
 	 *
 	 * @param PlaylistId the Playlist id
 	 * @return the Playlists by id
-	 * @throws ResourceNotFoundException the resource not found exception
+	 * @throws GlobalHttpException the resource not found exception
 	 */
 	@GetMapping("/playlists/{id}")
-	public ResponseEntity<Playlist> getPlaylistsById(@PathVariable(value = "id") Long playlistId) throws ResourceNotFoundException {
+	public ResponseEntity<Playlist> getPlaylistsById(@PathVariable(value = "id") Long playlistId) throws GlobalHttpException {		
+		User userLogged = getUserLogged();
+		
 		Playlist playlist = playlistRepository
 			  				.findById(playlistId)
-	  						.orElseThrow(() -> new ResourceNotFoundException(HttpStatus.NOT_FOUND, "Playlist not found with id : " + playlistId));
-	  
-		return ResponseEntity.ok().body(playlist);
+	  						.orElseThrow(() -> new GlobalHttpException(HttpStatus.NOT_FOUND, "Playlist not found with id : " + playlistId));
+		
+		if (playlist.getUser().getId() == userLogged.getId()) {			
+			return ResponseEntity.ok().body(playlist);
+		}
+		
+		throw new GlobalHttpException(HttpStatus.FORBIDDEN, "Not authorized"); 
 	}
 
 	/**
@@ -73,15 +95,13 @@ public class PlaylistController {
 	 */
 	@PostMapping("/playlists")
 	public Playlist createPlaylist(@Validated @RequestBody PlaylistParams params) {
-		User user = userRepository
-			  				.findById(params.getUser_id())
-	  						.orElseThrow(() -> new ResourceNotFoundException(HttpStatus.NOT_FOUND, "User not found with id : " + params.getUser_id()));
+		User userLogged = getUserLogged();
 		
 		params.setCreatedAt(new Date());
 		params.setUpdatedAt(new Date());
 		
 		Playlist playlist = new Playlist(params);
-		playlist.setUser(user);
+		playlist.setUser(userLogged);
 		
 		if (params.getTitle_ids() != null) {
 	    	List<Title> titles = titleRepository.findAllById(params.getTitle_ids());
@@ -97,27 +117,34 @@ public class PlaylistController {
 	 * @param PlaylistId the Playlist id
 	 * @param PlaylistDetails the Playlist details
 	 * @return the response entity
-	 * @throws ResourceNotFoundException the resource not found exception
+	 * @throws GlobalHttpException the resource not found exception
 	 */
 	@PutMapping("/playlists/{id}")
 	public ResponseEntity<Playlist> updatePlaylist(@PathVariable(value = "id") Long playlistId, @RequestBody PlaylistParams playlistDetails)
-			throws ResourceNotFoundException {
-	    Playlist playlist = playlistRepository
-	            			.findById(playlistId)
-	            			.orElseThrow(() -> new ResourceNotFoundException(HttpStatus.NOT_FOUND, "Playlist not found with id : " + playlistId));
-
-	    if (playlistDetails.getName() != null) {
-	    	playlist.setName(playlistDetails.getName());
-	    }
-	    
-	    if (playlistDetails.getTitle_ids() != null) {
-	    	List<Title> titles = titleRepository.findAllById(playlistDetails.getTitle_ids());
-	    	playlist.setTitles(new HashSet<Title>(titles));
-	    }
-	    
-	    playlist.setUpdatedAt(new Date());
-	    final Playlist updatedPlaylist = playlistRepository.save(playlist);
-	    return ResponseEntity.ok(updatedPlaylist);
+			throws GlobalHttpException {
+		
+		User userLogged = getUserLogged();
+		
+		Playlist playlist = playlistRepository
+			  				.findById(playlistId)
+	  						.orElseThrow(() -> new GlobalHttpException(HttpStatus.NOT_FOUND, "Playlist not found with id : " + playlistId));
+		
+		if (playlist.getUser().getId() == userLogged.getId()) {			
+		    if (playlistDetails.getName() != null) {
+		    	playlist.setName(playlistDetails.getName());
+		    }
+		    
+		    if (playlistDetails.getTitle_ids() != null) {
+		    	List<Title> titles = titleRepository.findAllById(playlistDetails.getTitle_ids());
+		    	playlist.setTitles(new HashSet<Title>(titles));
+		    }
+		    
+		    playlist.setUpdatedAt(new Date());
+		    final Playlist updatedPlaylist = playlistRepository.save(playlist);
+		    return ResponseEntity.ok(updatedPlaylist);
+		}
+		
+		throw new GlobalHttpException(HttpStatus.FORBIDDEN, "Not authorized");
 	}
 	
 
@@ -130,13 +157,19 @@ public class PlaylistController {
 	 */
 	@DeleteMapping("/playlists/{id}")
 	public Map<String, Boolean> deletePlaylist(@PathVariable(value = "id") Long playlistId) throws Exception {
-	    Playlist playlist = playlistRepository
-	            			.findById(playlistId)
-	            			.orElseThrow(() -> new ResourceNotFoundException(HttpStatus.NOT_FOUND, "Playlist not found with id : " + playlistId));
-
-	    playlistRepository.delete(playlist);
-	    Map<String, Boolean> response = new HashMap<>();
-	    response.put("deleted", Boolean.TRUE);
-	    return response;
+		User userLogged = getUserLogged();
+		
+		Playlist playlist = playlistRepository
+			  				.findById(playlistId)
+	  						.orElseThrow(() -> new GlobalHttpException(HttpStatus.NOT_FOUND, "Playlist not found with id : " + playlistId));
+		
+		if (playlist.getUser().getId() == userLogged.getId()) {			
+		    playlistRepository.delete(playlist);
+		    Map<String, Boolean> response = new HashMap<>();
+		    response.put("deleted", Boolean.TRUE);
+		    return response;
+		}
+		
+		throw new GlobalHttpException(HttpStatus.FORBIDDEN, "Not authorized");
 	}
 }
